requirements.txt
torch~=2.5.1
torchvision~=0.20.1
timm~=0.4.12
librosa~=0.10.2.post1
numpy~=1.26.4
pandas~=2.2.3
scipy~=1.13.1
transformers~=4.48.3
einops~=0.8.1
termcolor~=2.4.0
wandb~=0.17.9
PyYAML~=6.0.2
yacs~=0.1.8
fire~=0.7.0
tqdm~=4.49.0
matplotlib~=3.9.4
seaborn~=0.13.2
python-dotenv~=1.0.1